routerをauth, habitsで分離
